MaisWebapp/cnn_demo/requirements.txt
Flask==1.0.2
matplotlib==2.2.3
numpy==1.15.1
Pillow==6.2.0
torchvision==0.4.1
pytorch-pretrained-bert==0.6.2
tensorflow==2.0.0
